Create das5.md
I'm not entirely sure whether this fits the intended purposes of the NLeSC guide, because this text is very specific and very 'hands-on'. Then again, the DAS is used by many at the NLeSC and a lot. This text is intended for getting new employees up to speed. But if it doesn't fit the guide, I can fully understand.