reorganize chapters: add technology guides section
Also moved language guides to the root folder level, because I find it annoying to have to click into the best_practices folder.

Note: datasets still has to be moved to technologies, but there is an outstanding PR on that and I don't want to stress git's ability to detect moved files too much, otherwise we lose the ability to easily browse edit history.