update PR link checker workflow
Use a different changed-files checker (the old one was no longer
maintained) and also make the link checking more fine-grained by only
checking added lines for broken links.

We now use a matrix to parallelize over changed files. This slows down
the complete workflow a bit in some cases, but it does offer a nicer
user experience, because the files that are checked will now be shown in
the GitHub UI itself, instead of only in the workflow run logs.

Lychee is installed directly from the GitHub nightly release. This is
really fast. The original reason, though, was because there is no Ubuntu
package for Lychee. Works out well, though.

Determining the base SHA is a bit convoluted, because GitHub does not
provide this in a usable way. See
https://www.kenmuse.com/blog/the-many-shas-of-a-github-pull-request/ for
a good overview of this problem. The git rev-parse command was suggested
there. The fetch depth must be at least two, otherwise the base SHA
cannot be used.