[CI] link checker: update action version + ignore 429
Especially gnu.org is giving a lot of 429 statuses (too many requests), seems like it doesn't like GitHub spam. There doesn't seem to be much one can do about this, so let's just ignore this code (add it to the accept list) and hope people run the link checker on the GNU links locally every now and then. In a way, accepting this code is not wrong, because a benevolent server will not give 429 if the URL is invalid (it will give 404 instead), so in a way 429 may signify that the URL is valid (it just won't load for you right now).